refactor: Create separate django application for home page
