+++ server/apps/main/urls.py
@@ -3,8 +3,8 @@ from django.urls import path
 from .views import show_main_page
 
 
-app_name = 'core'
+app_name = 'main'
 
 urlpatterns = (
-    path('', show_main_page, name='profile'),
+    path('', show_main_page, name='home'),
 )
