+++ server/settings/components/common.py
@@ -14,6 +14,7 @@ INSTALLED_APPS: tuple[str, ...] = (
     # Internal apps go here:
     'server.apps.accounts',
     'server.apps.core',
+    'server.apps.main',
     'server.apps.ui_client',
 
     # Default django apps:
