+++ server/urls.py
@@ -11,8 +11,8 @@ from django.utils.translation import gettext_lazy as _
 from django.views.generic import TemplateView
 from health_check import urls as health_urls
 from server.apps.accounts import urls as accounts_urls
-from server.apps.core import urls as learning_urls
-from server.apps.core.views import show_main_page
+from server.apps.main import urls as main_urls
+from server.apps.main.views import show_main_page
 
 
 # Customizing the admin panel via a global admin site object
@@ -36,7 +36,7 @@ __humans_view = TemplateView.as_view(
 
 urlpatterns = (
     path('accounts/', include(accounts_urls, namespace='accounts')),
-    path('learning/', include(learning_urls, namespace='core')),
+    path('main/', include(main_urls, namespace='main')),
     path('health/', include(health_urls)),
     path('admin/', admin.site.urls),
     path('robots.txt', __robots_view),
