+++ tests/test_apps/test_main/test_views/test_home_view.py
@@ -6,18 +6,18 @@ from django.urls import reverse
 
 
 def test_profile_unauthorized(client: Client) -> None:
-    '''This test ensures that profile page requires auth.'''
-    response = client.get(reverse('core:profile'))
+    '''This test ensures that home page requires auth.'''
+    response = client.get(reverse('main:home'))
     assert response.status_code == HTTPStatus.FOUND
 
 
 def test_profile_authorized(client: Client, django_user_model: User) -> None:
-    '''This test ensures that profile page is accessible.'''
+    '''This test ensures that home page is accessible.'''
 
     username, password = 'test-user1', 'test-user1'
     test_user = django_user_model.objects.create_user(
         username=username, password=password
     )
     client.force_login(test_user)
-    response = client.get(reverse('core:profile'))
+    response = client.get(reverse('main:home'))
     assert response.status_code == HTTPStatus.OK
